% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfToMap.R
\name{dfToMap}
\alias{dfToMap}
\title{Data Frame to Map}
\usage{
dfToMap(map.df, fill = NULL)
}
\arguments{
\item{map.df}{The directory containing the shapefile}

\item{fill}{The column to be used as the fill}
}
\value{
A ggplot2 object
}
\description{
\code{dfToMap} plots dataframes of the class \code{map.df} using \code{ggplot2}.
}
\details{
See vignette for examples on how to use this function.
}
