% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fredMaps.R
\name{fredMaps}
\alias{fredMaps}
\title{FredMaps}
\usage{
fredMaps(data, x, state_col, low = "yellow", high = "red",
  states = NULL)
}
\arguments{
\item{data}{A dataframe containing a variable to plot and a variable of state names}

\item{x}{A variable in \code{data} to plot}

\item{state_col}{The variable in \code{data} of state names}

\item{low}{Color of lower values of \code{x}}

\item{high}{Color of higher values of \code{x}}

\item{states}{If specified, plots list of states provided}
}
\value{
A ggplot2 object
}
\description{
\code{fredMaps} creates a chloropleth map of the continental United States.
The color and intensity of the fill are based on the values of a user specified variable.
}
\examples{
fredMaps(data = obese, x = AdultsObese, state_col = State)
fredMaps(obese, AdultsObese, State, low = "blue", high = "purple",
states = c("illinois", "indiana", "iowa", "kentucky", "michigan", "minnesota","missouri", "north dakota", "ohio", "south dakota", "wisconsin"))
fredMaps(data = college_enrollment, x = undergrad, state_col = state)
}
