% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fredMapsPoint.R
\name{fredMapsPoint}
\alias{fredMapsPoint}
\title{FredMapsPoint}
\usage{
fredMapsPoint(data, size_var, color_var, long, lat)
}
\arguments{
\item{data}{A dataframe containing two variables to plot and a variable of state names}

\item{size_var}{A variable in \code{data} which determines size of points}

\item{color_var}{A variable in \code{data} which determines color of points}

\item{long}{The variable in \code{data} containing longitude values}

\item{lat}{The variable in \code{data} containing latitude values}
}
\value{
A ggplot2 object
}
\description{
\code{fredMapsPoint} creates a chloropleth map of the continental United States.
The color and size of the points are based on the values of two user specified variables.
}
\examples{
fredMapsPoint(college_enrollment, size=enrollment, color=label, long=long, lat=lat)
fredMapsPoint(college_enrollment, size=enrollment, color=undergrad, long=long, lat=lat)
}
