% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdp.R
\docType{data}
\name{gdp}
\alias{gdp}
\title{State GDP Per Capita 2015 Data}
\format{A data frame with 51 rows and two variables.
The variables are as follows:
\describe{
   \item{X}{state}
   \item{gdpPerCap}{Nominal GDP per capita}
}}
\source{
The data is a subset of the \href{https://data.world/gmoney/gdp-per-capita-by-state}{GDP per Capita}
dataset for the years 1997 - 2015.
}
\usage{
gdp
}
\description{
This data frame contains nominal GDP per capita by state for 2015. The
District of Columbia is included in the dataset.
}
\examples{
summary(gdp)
}
\keyword{datasets}
