% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeOnState.R
\name{mergeOnState}
\alias{mergeOnState}
\title{Merge on State Function}
\usage{
mergeOnState(df1, col1, df2, col2, merge = "left")
}
\arguments{
\item{df1}{A dataframe containing a column with state names}

\item{col1}{The column within \code{df1} containing state names}

\item{df2}{A dataframe containing a column with state names}

\item{col2}{The column within \code{df2} containing state names}

\item{merge}{The type of merge ("left", "right", "inner", "full") that should
be used to merge the two dataframes}
}
\value{
A dataframe
}
\description{
\code{mergeOnState} merges two dataframe on state names. The columns
containing state names in the two dataframes do not have to have the same format.
}
\examples{
data(bingedrinking)
data(gdp)
mergeOnState(bingedrinking, LocationAbbr, gdp, X, merge = "inner")
}
