% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properCase.R
\name{properCase}
\alias{properCase}
\title{Proper Case Function}
\usage{
properCase(x)
}
\arguments{
\item{x}{A string}
}
\value{
A string with the first character capitalized
}
\description{
\code{properCase} makes the first character in a word uppercase and all
other characters lowercase.
}
\details{
This function capitalizes the first character is a word and makes all subsequent
characters lowercase.  This capitalization sequence is often known as proper case.
}
\examples{
properCase("hello")
properCase("HELLO")
properCase("hello world")
properCase("HelLO WorLD")
}
