% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapeToDF.R
\name{shapeToDF}
\alias{shapeToDF}
\title{Shape File to Data Frame}
\usage{
shapeToDF(dir, file)
}
\arguments{
\item{dir}{The directory containing the shapefile}

\item{file}{The name of the shapefile without extension}
}
\value{
A map.df object
}
\description{
\code{shapeToDF} reads in a shapefile and converts it into a
dataframe to allow for easy plotting in \code{ggplot2}.

\code{shapeToDF}
converts a \code{SpatialPolygonsDataFrame} to a \code{data.frame}, however is not built to
handle a \code{SpatialLinesDataFrame} or a \code{SpatialPointaDataFrame}.
}
\details{
See vignette for examples on how to use this function.
}
