% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapeToMap.R
\name{shapeToMap}
\alias{shapeToMap}
\title{Shape File to Map}
\usage{
shapeToMap(dir, file)
}
\arguments{
\item{dir}{The directory containing the shapefile}

\item{file}{The name of the shapefile without extension}
}
\value{
A ggplot2 object
}
\description{
\code{shapeToMap} reads in a shapefile, converts it to a dataframe
and plots the resulting dataframe using \code{ggplot2}.

\code{shapeToMap}
converts a \code{SpatialPolygonsDataFrame} to a \code{data.frame}, however is not built to
handle a \code{SpatialLinesDataFrame} or a \code{SpatialPointaDataFrame}.
}
\details{
See vignette for examples on how to use this function.
}
