% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stateNames.R
\name{stateNames}
\alias{stateNames}
\title{State Names Function}
\usage{
stateNames(state, tolower = TRUE, abb = FALSE, replace.na = FALSE)
}
\arguments{
\item{state}{a string containing a state name or abbreviation}

\item{tolower}{TRUE if state names should be outputted in lowercase,
FALSE if state names should be outputted in proper case}

\item{abb}{TRUE if state abbreviation is to be returned, FALSE if full name is to be returned}

\item{replace.na}{TRUE if non-states are to be replaced with \code{NA}s,
FALSE if non-states are to be returned unchanged}
}
\value{
a named character containing normalized state name(s) or abbreviation(s).
}
\description{
\code{stateNames} converts U.S. state names to a user specified normalized
format.  By default, \code{stateNames} outputs results as lowercase full state names. The user
can specify if the output should be full names or abbreviation and lowercase or proper case
(first letter in the word capitalized).  If a string is not one of the 50 U.S. states,
that string will be returned unchanged, however, the
user can specifiy if it should be replaced with an \code{NA}.
}
\examples{
stateNames("NY")
stateNames("new york", abb = TRUE)
stateNames(c("New york", "ca", "ct"), tolower = FALSE)
stateNames(c("New york", "ca", "ct"), tolower = FALSE, abb = TRUE)
stateNames(c("New york", "Boston", "ct"), tolower = FALSE, abb = TRUE, replace.na = TRUE)
}
